Added some futher options and comments for clarity
